+++ src/main/java/com/getgreetapp/greetapp/web/rest/GangUserResource.java
@@ -1,16 +1,25 @@
 package com.getgreetapp.greetapp.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
+import com.getgreetapp.greetapp.domain.Gang;
 import com.getgreetapp.greetapp.domain.GangUser;
 import com.getgreetapp.greetapp.domain.User;
+import com.getgreetapp.greetapp.repository.GangRepository;
 import com.getgreetapp.greetapp.repository.GangUserRepository;
 import com.getgreetapp.greetapp.repository.UserRepository;
+import com.getgreetapp.greetapp.security.SecurityUtils;
+import com.getgreetapp.greetapp.specification.rules.CanCreateGangUser;
+import com.getgreetapp.greetapp.specification.rules.CanViewGang;
+import com.getgreetapp.greetapp.specification.rules.IsAdminOfGang;
 import com.getgreetapp.greetapp.web.rest.errors.BadRequestAlertException;
 import com.getgreetapp.greetapp.web.rest.util.HeaderUtil;
 import io.github.jhipster.web.util.ResponseUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.parameters.P;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -33,8 +42,17 @@ public class GangUserResource {
 
     private final GangUserRepository gangUserRepository;
 
-    public GangUserResource(GangUserRepository gangUserRepository) {
+    private final UserRepository userRepository;
+
+    private final GangRepository gangRepository;
+
+    public GangUserResource(
+        GangUserRepository gangUserRepository,
+        UserRepository userRepository,
+        GangRepository gangRepository) {
         this.gangUserRepository = gangUserRepository;
+        this.userRepository = userRepository;
+        this.gangRepository = gangRepository;
     }
 
     /**
@@ -51,10 +69,17 @@ public class GangUserResource {
         if (gangUser.getId() != null) {
             throw new BadRequestAlertException("A new gangUser cannot already have an ID", ENTITY_NAME, "idexists");
         }
-        GangUser result = gangUserRepository.save(gangUser);
-        return ResponseEntity.created(new URI("/api/gang-users/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
-            .body(result);
+
+        CanCreateGangUser canCreateGangUser = new CanCreateGangUser(this.userRepository);
+
+        if (canCreateGangUser.isSatisfiedBy(gangUser)) {
+            GangUser result = gangUserRepository.save(gangUser);
+            return ResponseEntity.created(new URI("/api/gang-users/" + result.getId()))
+                .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
+                .body(result);
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     /**
@@ -73,10 +98,17 @@ public class GangUserResource {
         if (gangUser.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
-        GangUser result = gangUserRepository.save(gangUser);
-        return ResponseEntity.ok()
-            .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, gangUser.getId().toString()))
-            .body(result);
+
+        CanCreateGangUser canCreateGangUser = new CanCreateGangUser(this.userRepository);
+
+        if (canCreateGangUser.isSatisfiedBy(gangUser)) {
+            GangUser result = gangUserRepository.save(gangUser);
+            return ResponseEntity.ok()
+                .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, gangUser.getId().toString()))
+                .body(result);
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     /**
@@ -86,6 +118,7 @@ public class GangUserResource {
      */
     @GetMapping("/gang-users")
     @Timed
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
     public List<GangUser> getAllGangUsers() {
         log.debug("REST request to get all GangUsers");
         return gangUserRepository.findAll();
@@ -98,9 +131,18 @@ public class GangUserResource {
      */
     @GetMapping("/gang-users-by-group/{groupId}")
     @Timed
-    public List<GangUser> getAllGangUsers(@PathVariable Long groupId) {
+    public Object getAllGangUsers(@PathVariable Long groupId) {
         log.debug("REST request to get all GangUsers for group");
-        return gangUserRepository.findByGroup(groupId);
+
+        CanViewGang canViewGang = new CanViewGang(this.userRepository);
+        Optional<Gang> optionalGang = this.gangRepository.findById(groupId);
+        Gang gang = optionalGang.get();
+
+        if (canViewGang.isSatisfiedBy(gang)) {
+            return gangUserRepository.findByGroup(groupId);
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     /**
@@ -113,8 +155,16 @@ public class GangUserResource {
     @Timed
     public ResponseEntity<GangUser> getGangUser(@PathVariable Long id) {
         log.debug("REST request to get GangUser : {}", id);
-        Optional<GangUser> gangUser = gangUserRepository.findById(id);
-        return ResponseUtil.wrapOrNotFound(gangUser);
+        Optional<GangUser> optionalGangUser = gangUserRepository.findById(id);
+        GangUser gangUser = optionalGangUser.get();
+
+        CanViewGang canViewGang = new CanViewGang(this.userRepository);
+
+        if (canViewGang.isSatisfiedBy(gangUser.getGang())) {
+            return ResponseUtil.wrapOrNotFound(optionalGangUser);
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     /**
@@ -128,7 +178,18 @@ public class GangUserResource {
     public ResponseEntity<Void> deleteGangUser(@PathVariable Long id) {
         log.debug("REST request to delete GangUser : {}", id);
 
-        gangUserRepository.deleteById(id);
-        return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        Optional<GangUser> optionalGangUser = gangUserRepository.findById(id);
+        GangUser gangUser = optionalGangUser.get();
+
+        Optional<String> login = SecurityUtils.getCurrentUserLogin();
+        User loggedInUser = userRepository.findOneByLogin(login.get()).get();
+
+        IsAdminOfGang isAdminOfGang = new IsAdminOfGang(loggedInUser);
+        if (isAdminOfGang.isSatisfiedBy(gangUser.getGang())) {
+            gangUserRepository.deleteById(id);
+            return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 }
